disable propTypes check for StandardFiltersHOC
